Blinkenlight server for PiDP8 (in 11_pidp8): LED anti-flicker and "slow" lightbulb behaviour,
blinkenlightapitest now with increment, decrement and shift pattern tests.

All: "u_int64_t" replaced with "uint64_t", less compiler warnings.

+++ projects/02.3_simh.4.x.jh/src/REALCONS/realcons_simh.c
@@ -127,7 +127,7 @@ t_stat realcons_simh_set_service_interval(int32 flg, char *cptr)
 		return SCPE_ARG;
 	if (*tptr != 0)
 		return SCPE_ARG;
-	cpu_realcons->service_interval_msec = val;
+	cpu_realcons->service_interval_msec = (unsigned)val;
 	return SCPE_OK;
 }
